core. mems parameters read and set. android libneurosdk. mems channel and params wrappers

+++ core/include/channels/mems_data.h
@@ -5,15 +5,15 @@ namespace Neuro {
 
 struct MEMS{
     struct Accelerometer{
-        short X;
-        short Y;
-        short Z;
+        double X;
+        double Y;
+        double Z;
     } accelerometer;
 
     struct Gyroscope{
-        short X;
-        short Y;
-        short Z;
+        double X;
+        double Y;
+        double Z;
     } gyroscope;
 };
